timer: use cached current time to arm the timer
Otherwise it's possible to modify current mocked time after call to
timer.Reset but *before* m.wait() function actually executes. This leads
to timer being armed using *new* current time, instead of the one that
was set at the time Reset() was called.

This leads to deadlock in tests that execute in multiple goroutines.

Signed-off-by: Pavel Borzenkov <pavel.borzenkov@gmail.com>